Make Sup's singleton part of the Redwood module
