+++ lib/sup/account.rb
@@ -25,7 +25,7 @@ class Account < Person
 end
 
 class AccountManager
-  include SupSingleton
+  include Singleton
 
   attr_accessor :default_account
