+++ lib/sup/index.rb
@@ -55,7 +55,7 @@ EOS
     def method_missing m; @h[m.to_s] end
   end
 
-  include SupSingleton
+  include Singleton
 
   def initialize dir=BASE_DIR
     @dir = dir
