+++ lib/sup/poll.rb
@@ -3,7 +3,7 @@ require 'thread'
 module Redwood
 
 class PollManager
-  include SupSingleton
+  include Singleton
 
   HookManager.register "before-add-message", <<EOS
 Executes immediately before a message is added to the index.
