+++ lib/sup/source.rb
@@ -187,7 +187,7 @@ module SerializeLabelsNicely
 end
 
 class SourceManager
-  include SupSingleton
+  include Singleton
 
   def initialize
     @sources = {}
